Implemented day 09 part 1. Also refactored helper classes and their callers to use long longs.

+++ helper-classes/list-reader.h
@@ -12,7 +12,7 @@ namespace data {
 
 // Read in data from file and return as a line-by-line list.
 std::vector<std::string> GetStringList(const std::string& filename);
-std::vector<int> GetIntList(const std::string& filename);
+std::vector<long long int> GetIntList(const std::string& filename);
 
 } // namespace data
